+++ CHANGELOG.md
@@ -1,2 +1,14 @@
+# 0.2.0
+## Added
+- Documentation for all public items
+- `AppId`
+- `SteamClient::get_game_library_capsule`
+- `SteamClient::resolve_vanity_url`
+- `SteamClient::get_recent_games`
+## Changed
+- Removed caching of vanity urls
+- All calls to `SteamClient` now need player id instead of vanity/player id
+- Moved some items around, renamed some member values
+
 # 0.1.0
 ### Initial release